+++ unfucker.js
@@ -1,1 +1,0 @@
-//Project has a new maintainer! Go to https://github.com/ClangPan/dashboard-unfucker, select the version you want, and hit the "raw" button with tampermonkey installed to install the script